Add: 댓글기능  1차 추가 및 main페이지 컴포넌트로 분리
