+++ src/pages/kanghoYoo/Components/Footer.js
@@ -1,0 +1,11 @@
+export default function Footer() {
+  return (
+    <footer className="Footer">
+      <p>
+        Westargram 정보 지원 홍보센터 API 채용정보 개인정보정보처리방침 약관
+        디렉터리 프로필 해시태그 언어
+      </p>
+      <span>&copy; 2022 WESTARGRAM</span>
+    </footer>
+  );
+}
