--- src/pages/kanghoYoo/Components/commentData.js
@@ -1,22 +1,0 @@
-const COMMENT_LIST = [
-  {
-    id: 1,
-    userName: 'wecode',
-    content: 'Welcome to world best coding bootcamp!',
-    isLiked: true,
-  },
-  {
-    id: 2,
-    userName: 'joonsikyang',
-    content: 'Hi there.',
-    isLiked: false,
-  },
-  {
-    id: 3,
-    userName: 'jayPark',
-    content: 'Hey.',
-    isLiked: false,
-  },
-];
-
-export default COMMENT_LIST;
