+++ src/pages/kanghoYoo/Main/MainYoo.js
@@ -1,281 +1,26 @@
-import React from 'react';
+// import React, { useEffect } from 'react';
 import './MainYoo.scss';
-import { Link } from 'react-router-dom';
 import Nav from '../../../components/Nav/Nav';
-import { useState } from 'react';
-import commentList from '.././Components/CommentList';
+import Main from '../Components/Main';
+import Aside from '../Components/Aside';
+import Footer from '../Components/Footer';
 
 function MainYoo() {
-  const commentValue = [
-    { name: 'vertex7785', text: '시간이 되게 빠르게 지나가네...' },
-    {
-      name: 'coke_0kal',
-      text: '저기 오른쪽으로 가시면 극장이 있죠. 극장에서 쭉 가시면 시장이 나옵니다. 시장에서 왼쪽으로 돌면 편의점, 편의점 건녀편에 이발소, 이발소 옆 골목으로 들어가서 오른쪽 첫번째 집이 저희 집입니다.',
-    },
-  ];
-
-  const [comment, setComment] = useState([]);
-  // const [arrComment, setArrComment] = useState([]);
-  const upload = document.getElementsByClassName('.comments')[0];
-
-  function uploads() {
-    commentValue.map(el => {
-      return (
-        <li className="mentsWrap">
-          <span className="commentId">{el.name}</span> {el.text}
-        </li>
-      );
-    });
-  }
-
-  function commentInput(e) {
-    setComment(e.target.value);
-  }
-
-  const cmtInput = document.getElementsByClassName('.commentInput')[0];
-
-  function commentSend() {
-    if (cmtInput.value !== '') {
-    }
-  }
-
-  function enter(e) {
-    if (e.key === 'enter') {
-      commentSend();
-    }
-  }
+  // useEffect(() => {
+  //   fetch('http://localhost:3000/data/commentData.json')
+  //     .then(response => response.json())
+  //     .then(data => {
+  //       setComment(data);
+  //     });
+  // }, []);
 
   return (
     <>
       <Nav />
-      <main className="main">
-        <div className="feeds">
-          <article className="article">
-            <div className="articleTop">
-              <div className="account">
-                <div>
-                  <img src="./profile.png" className="profile" />
-                </div>
-                <div className="name">
-                  <span>Kangho_Yoo</span>
-                </div>
-              </div>
-              <div>
-                <i className="fas fa-ellipsis-h" />
-              </div>
-            </div>
-            <div>
-              <img
-                src="https://images.unsplash.com/photo-1504297050568-910d24c426d3?ixlib=rb-1.2.1&ixid=MnwxMjA3fDF8MHxlZGl0b3JpYWwtZmVlZHw0OXx8fGVufDB8fHx8&auto=format&fit=crop&w=800&q=60"
-                alt="메인 이미지"
-                className="artImg"
-              />
-            </div>
-            <div className="iconsWrap">
-              <div className="icons">
-                <button className="heart" onclick="toggleImg">
-                  <i className="fas fa-heart" />
-                </button>
-                <i className="far fa-comment" />
-                <i className="fas fa-file-upload" />
-              </div>
-              <div>
-                <i className="fas fa-bookmark" />
-              </div>
-            </div>
-            <div className="alert">
-              <img
-                src="https://images.unsplash.com/photo-1644329901564-1e47f5d1b81a?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHw2OHx8fGVufDB8fHx8&auto=format&fit=crop&w=800&q=60"
-                alt="이미지"
-                className="loveProfile"
-              />
-              <span className="good">
-                <span className="bold">alneworld</span>님 외
-                <span className="bold">10명</span>이좋아합니다
-              </span>
-            </div>
-            <div className="comments">
-              {commentValue.map(el => {
-                return (
-                  <li className="mentsWrap">
-                    <span className="commentId">{el.name}</span> {el.text}
-                  </li>
-                );
-              })}
-              {/* {commentList.map((item, i) => {
-                return (
-                  <div className="commentWrap" key={i} id={id} text="item">
-                    <li className="comment" key={props.key}>
-                      <span className="commentId">{props.id}</span> {props.text}
-                    </li>
-                  </div>
-                );
-              })} */}
-            </div>
-            <div className="comments" />
-            <div>
-              <p className="mentTime">42분 전</p>
-            </div>
-
-            <div className="input">
-              <input
-                type="text"
-                className="commentInput"
-                value={comment}
-                onChange={commentInput}
-                placeholder="댓글 달기..."
-                onProgress={enter}
-              />
-              <input
-                type="button"
-                className="send"
-                onclick={commentSend}
-                value="게시"
-              />
-            </div>
-          </article>
-        </div>
-      </main>
+      <Main />
       <div className="mainRight">
-        <div className="right1">
-          <div className="anotherProfile">
-            <img
-              src="https://media.vlpt.us/images/inyong_pang/post/f0ea605d-c2d9-460c-aedc-a0ec77e6759f/wecode.png"
-              alt="다른 프로필"
-              className="profiles"
-            />
-          </div>
-          <div className="anotherInfo">
-            <p id="another-name">wecode_bootcamp</p>
-            <span id="en-name" className="names">
-              WeCode
-            </span>
-            <span id="ko-name" className="names">
-              위코드
-            </span>
-          </div>
-        </div>
-        <div className="right2">
-          <div className="storyTop">
-            <span className="top1">스토리</span>
-            <span className="top2">모두 보기</span>
-          </div>
-          <div>
-            <div className="stories">
-              <div className="anotherProfile">
-                <img
-                  src="https://images.unsplash.com/photo-1499996860823-5214fcc65f8f?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8NHx8aHVtYW58ZW58MHx8MHx8&auto=format&fit=crop&w=800&q=60"
-                  alt="다른 프로필"
-                  className="anotherImg"
-                />
-              </div>
-              <div className="anotherInfo">
-                <p>kali_s</p>
-                <span className="times">30분 전</span>
-              </div>
-            </div>
-            <div className="stories">
-              <div className="anotherProfile">
-                <img
-                  src="https://images.unsplash.com/photo-1511367461989-f85a21fda167?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8NXx8aHVtYW58ZW58MHx8MHx8&auto=format&fit=crop&w=800&q=60"
-                  alt="다른 프로필"
-                  className="anotherImg"
-                />
-              </div>
-              <div className="anotherInfo">
-                <p>yum_s</p>
-                <span className="times">16분 전</span>
-              </div>
-            </div>
-            <div className="stories">
-              <div className="anotherProfile">
-                <img
-                  src="https://images.unsplash.com/photo-1593529467220-9d721ceb9a78?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MTl8fGh1bWFufGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=800&q=60"
-                  alt="다른 프로필"
-                  className="anotherImg"
-                />
-              </div>
-              <div className="anotherInfo">
-                <p>natali_A</p>
-                <span className="times">14분 전</span>
-              </div>
-            </div>
-            <div className="stories">
-              <div className="anotherProfile">
-                <img
-                  src="https://images.unsplash.com/photo-1539640038819-57bb520c7d2d?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MjB8fGh1bWFufGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=800&q=60"
-                  alt="다른 프로필"
-                  className="anotherImg"
-                />
-              </div>
-              <div className="anotherInfo">
-                <p>cho_ha</p>
-                <span className="times">2시간 전</span>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className="right3">
-          <div className="storyTop">
-            <span className="top1">회원님을 위한 추천</span>
-            <span className="top2">모두 보기</span>
-          </div>
-          <div>
-            <div className="recomment">
-              <div className="anotherProfile">
-                <img
-                  src="https://images.unsplash.com/photo-1501196354995-cbb51c65aaea?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8Mzd8fCUwOHBlb3BsZXxlbnwwfHwwfHw%3D&auto=format&fit=crop&w=800&q=60"
-                  alt="다른 프로필"
-                  className="anotherImg2"
-                />
-              </div>
-              <div className="anotherRecomment">
-                <p>de_lo</p>
-                <span className="flowing">
-                  _blosanrang님 외 22명이 팔로잉 중
-                </span>
-              </div>
-              <Link to="/main-Yoo">팔로우</Link>
-            </div>
-            <div className="recomment">
-              <div className="anotherProfile">
-                <img
-                  src="https://images.unsplash.com/photo-1463453091185-61582044d556?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MTl8fCUwOHBlb3BsZXxlbnwwfHwwfHw%3D&auto=format&fit=crop&w=800&q=60"
-                  alt="다른 프로필"
-                  className="anotherImg2"
-                />
-              </div>
-              <div className="anotherRecomment">
-                <p>Caldit_q</p>
-                <span className="flowing">kukuro_P님 외 10명이 팔로잉 중</span>
-              </div>
-              <Link to="/main-Yoo">팔로우</Link>
-            </div>
-            <div className="recomment">
-              <div className="anotherProfile">
-                <img
-                  src="https://images.unsplash.com/photo-1438761681033-6461ffad8d80?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxzZWFyY2h8M3x8JTA4cGVvcGxlfGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=800&q=60"
-                  alt="다른 프로필"
-                  className="anotherImg2"
-                />
-              </div>
-              <div className="anotherRecomment">
-                <p>tcp_best</p>
-                <span className="flowing">
-                  js_lover_html님 외 2명이 팔로잉 중
-                </span>
-              </div>
-              <Link to="/main-Yoo">팔로우</Link>
-            </div>
-          </div>
-        </div>
-        <footer className="footer">
-          <p>
-            Westargram 정보 지원 홍보센터 API 채용정보 개인정보정보처리방침 약관
-            디렉터리 프로필 해시태그 언어
-          </p>
-          <span>&copy; 2022 WESTARGRAM</span>
-        </footer>
+        <Aside />
+        <Footer />
       </div>
     </>
   );
