+++ src/pages/kanghoYoo/Main/MainYoo.scss
@@ -318,7 +318,7 @@
   }
 }
 
-.footer {
+Footer {
   p {
     font-size: 14px;
     color: rgb(184, 184, 184);
